spot instance

+++ dynamiccluster/aws_manager.py
@@ -59,22 +59,49 @@ class AWSManager(CloudManager):
                 else:
                     log.exeception("userdata file does not exist, can't create VM, please check your config.")
                     return None
-                reservation = self.conn.run_instances(self.config['image_id'], key_name=self.config['key_name'], max_count=1, min_count=1, user_data=userdata_string, security_groups=self.config['security_groups'], instance_type=self.config['instance_type'], placement=self.config['availability_zone']) 
-                for server in reservation.instances:
-                    server.add_tag('Name', server_name)
-                    instance = Instance(server.id)
+                if "spot_bid" in self.config:
+                    #start spot
+                    timeout=300
+                    if "spot_timeout" in self.config:
+                        timeout=self.config["spot_timeout"]
+                    valid_until=(datetime.datetime.utcnow()+datetime.timedelta(0, timeout)).isoformat()
+                    req=self.conn.request_spot_instances(self.config['spot_bid'], self.config['image_id'], count=1, valid_until=valid_until, user_data=userdata_string, placement=self.config['availability_zone'], instance_type=self.config['instance_type'], key_name=self.config['key_name'], security_groups=self.config['security_groups'])
+                    log.debug("create spot request %s" % req)
+                    if len(req)==0:
+                        log.error("unable to create spot request")
+                        raise CloudNotAvailableException()
+                    self.conn.create_tags([req[0].id], {'Name':server_name})
+                    instance = Instance(None)
                     instance.instance_name=server_name
-                    instance.vcpu_number=get_aws_vcpu_num_by_instance_type(server.instance_type)
-                    instance.creation_time=unix_time(datetime.datetime.strptime(server.launch_time, "%Y-%m-%dT%H:%M:%S.%fZ"))
                     instance.key_name=self.config['key_name']
                     instance.flavor=self.config['instance_type']
                     instance.security_groups=self.config['security_groups']
                     instance.availability_zone=self.config['availability_zone']
                     instance.image_uuid=self.config['image_id']
                     instance.cloud_resource=self.name
-                    instance.state=self.get_state(server)
-                    log.debug("launched a new instance: %s"%instance)
+                    instance.state=Instance.Pending
+                    instance.spot_id=req[0].id
+                    instance.spot_state=req[0].state
+                    instance.spot_price=req[0].price
+                    log.debug("submitted a spot request: %s"%instance)
                     new_instances.append(instance)
+                else:
+                    reservation = self.conn.run_instances(self.config['image_id'], key_name=self.config['key_name'], max_count=1, min_count=1, user_data=userdata_string, security_groups=self.config['security_groups'], instance_type=self.config['instance_type'], placement=self.config['availability_zone']) 
+                    for server in reservation.instances:
+                        server.add_tag('Name', server_name)
+                        instance = Instance(server.id)
+                        instance.instance_name=server_name
+                        instance.vcpu_number=get_aws_vcpu_num_by_instance_type(server.instance_type)
+                        instance.creation_time=unix_time(datetime.datetime.strptime(server.launch_time, "%Y-%m-%dT%H:%M:%S.%fZ"))
+                        instance.key_name=self.config['key_name']
+                        instance.flavor=self.config['instance_type']
+                        instance.security_groups=self.config['security_groups']
+                        instance.availability_zone=self.config['availability_zone']
+                        instance.image_uuid=self.config['image_id']
+                        instance.cloud_resource=self.name
+                        instance.state=self.get_state(server)
+                        log.debug("launched a new instance: %s"%instance)
+                        new_instances.append(instance)
             except:
                 log.exception("Unable to boot a new instance.")
                 pass
@@ -83,36 +110,95 @@ class AWSManager(CloudManager):
         return new_instances
     
     def update(self, instance):
-        log.notice("Getting instance %s..." % (instance.uuid))
         server=None
-        try:
-            servers = self.conn.get_only_instances(filters={"instance-id": instance.uuid})
-            if len(servers)==0:
-                log.info("instance %s doesn't exist"%instance.uuid)
-                instance.state=Instance.Inexistent
-                return instance
-            server=servers[0]
-        except:
-            log.exception("Unable to get instance details.")
-            raise CloudNotAvailableException()
-        instance.state=self.get_state(server)
-        log.debug("instance %s; Cloud Status: %s; State: %s" % (server.id, server.state, instance.state))
-        instance.creation_time=unix_time(datetime.datetime.strptime(server.launch_time, "%Y-%m-%dT%H:%M:%S.%fZ"))
-        instance.vcpu_number=get_aws_vcpu_num_by_instance_type(server.instance_type)
-        if instance.state==Instance.Active:
-            instance.ip=server.ip_address
-            instance.public_dns_name=server.public_dns_name
-            instance.availability_zone=server.placement
-            instance.image_uuid=server.image_id
+        if instance.spot_id is not None and instance.uuid==None:
+            log.notice("Updating spot request %s..." % (instance.spot_id))
+            try:
+                requests = self.conn.get_all_spot_instance_requests(request_ids=[instance.spot_id])
+                if len(requests)==0 or requests[0].state in ["closed", "canceled", "failed"]:
+                    instance.state=WorkerNode.Inexistent
+                    return instance
+                else:
+                    instance.spot_state=requests[0].state
+                    if requests[0].state=="active":
+                        servers = self.conn.get_only_instances(filters={"instance-id": requests[0].instance_id})
+                        if len(servers)==0:
+                            log.debug("instance %s doesn't exist yet"%requests[0].instance_id)
+                        else:
+                            server=servers[0]
+                            log.debug("instance %s is created as spot instance"%requests[0].instance_id)
+                            tags=self.conn.get_all_tags(filters={"resource-id":requests[0].id, "tag-key":"Name"})
+                            if len(tags)>0 and tags[0].name=="Name":
+                                server.add_tag('Name', tags[0].value)
+                            instance.uuid=requests[0].instance_id
+                            instance.vcpu_number=get_aws_vcpu_num_by_instance_type(server.instance_type)
+                            instance.creation_time=unix_time(datetime.datetime.strptime(server.launch_time, "%Y-%m-%dT%H:%M:%S.%fZ"))
+            except:
+                log.exception("Unable to get spot request details.")
+                raise CloudNotAvailableException()
+        if instance.uuid is not None and server is None:
+            log.notice("Getting instance %s..." % (instance.uuid))
+            try:
+                servers = self.conn.get_only_instances(filters={"instance-id": instance.uuid})
+                if len(servers)==0:
+                    log.info("instance %s doesn't exist"%instance.uuid)
+                    instance.state=Instance.Inexistent
+                    return instance
+                server=servers[0]
+            except:
+                log.exception("Unable to get instance details.")
+                raise CloudNotAvailableException()
+        if server is not None:
+            instance.state=self.get_state(server)
+            log.debug("instance %s; Cloud Status: %s; State: %s" % (server.id, server.state, instance.state))
+            instance.creation_time=unix_time(datetime.datetime.strptime(server.launch_time, "%Y-%m-%dT%H:%M:%S.%fZ"))
+            instance.vcpu_number=get_aws_vcpu_num_by_instance_type(server.instance_type)
+            if instance.state==Instance.Active:
+                instance.ip=server.ip_address
+                instance.public_dns_name=server.public_dns_name
+                instance.availability_zone=server.placement
+                instance.image_uuid=server.image_id
         return instance
 
     def list(self):
         """ List all instances with the configured prefix """
         try:
-            servers = self.conn.get_only_instances(filters={"tag:Name":self.config['instance_name_prefix']+"-*"})
             instances = []
+            requests = self.conn.get_all_spot_instance_requests(filters={"tag:Name":self.config['instance_name_prefix']+"-*", "state":["open","active"]})
+            spot_ids=[]
+            for request in requests:
+                instance = Instance(None)
+                if request.instance_id is not None:
+                    instance.uuid=request.instance_id
+                tags=self.conn.get_all_tags(filters={"resource-id":request.id, "tag-key":"Name"})
+                if len(tags)>0 and tags[0].name=="Name":
+                    instance.instance_name=tags[0].value
+                instance.key_name=self.config['key_name']
+                instance.flavor=self.config['instance_type']
+                instance.security_groups=self.config['security_groups']
+                instance.availability_zone=self.config['availability_zone']
+                instance.image_uuid=self.config['image_id']
+                instance.cloud_resource=self.name
+                instance.state=Instance.Pending
+                instance.spot_id=request.id
+                instance.spot_state=request.state
+                instance.spot_price=request.price
+                instances.append(instance)
+                spot_ids.append(request.id)
+            servers = self.conn.get_only_instances(filters={"tag:Name":self.config['instance_name_prefix']+"-*"})
             for server in servers:
-                instance=Instance(server.id)
+                instance=None
+                if server.spot_instance_request_id is not None and server.spot_instance_request_id in spot_ids:
+                    log.debug("instance %s has been created in the past 2 seconds!" % server.id)
+                    ins=[i for i in instances if i.spot_id==server.spot_instance_request_id]
+                    if len(ins)==0:
+                        log.debug("...but it is not in the list... strange...")
+                    else:
+                        instance=ins[0]
+                        instance.uuid=server.id
+                else:
+                    instance=Instance(server.id)
+                    instances.append(instance)
                 instance.instance_name=server.tags['Name']
                 instance.creation_time=unix_time(datetime.datetime.strptime(server.launch_time, "%Y-%m-%dT%H:%M:%S.%fZ"))
                 instance.key_name=server.key_name
@@ -129,7 +215,6 @@ class AWSManager(CloudManager):
                     instance.ip=server.ip_address
                     instance.public_dns_name=server.public_dns_name
                 instance.vcpu_number=get_aws_vcpu_num_by_instance_type(server.instance_type)
-                instances.append(instance)
             return instances
         except:
             log.exception("Encounter an error when connecting to AWS.")
@@ -139,7 +224,10 @@ class AWSManager(CloudManager):
         """ Terminate an instance """
         log.debug("Destroying instance %s (ip=%s)"%(instance.uuid,instance.ip))
         try:
-            servers = self.conn.terminate_instances(instance_ids=[instance.uuid])
+            if instance.uuid is not None:
+                self.conn.terminate_instances(instance_ids=[instance.uuid])
+            if instance.spot_id is not None:
+                self.conn.cancel_spot_instance_requests([instance.spot_id])
             #vm.delete_time = datetime.datetime.utcnow()
         #except NotFound as ex:
         #    log.debug("instance %s is already shut down"%(instance.uuid))
